backend/requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
sqlalchemy==2.0.43
pymysql==1.1.1
pydantic==2.11.7
pydantic-settings==2.10.1
python-dotenv==1.1.1
scikit-learn==1.7.1
pandas==2.3.1
numpy==2.3.2
shap==0.48.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
email-validator==2.2.0
python-multipart==0.0.20
